server/requirements.txt
tensorflow
pillow
flask
apscheduler
scipy
itsdangerous==2.1.2